EXP-56480 - Run a predefined command after dependency update
Extract GitDependenciesRepository into a new file
Use ChangeDir to navigate between folders
