+++ git/utilities.py
@@ -1,0 +1,14 @@
+import os
+import sys
+import subprocess
+
+class ChangeDir:
+    def __init__(self, newPath):
+        self.newPath = os.path.expanduser(newPath)
+
+    def __enter__(self):
+        self.savedPath = os.getcwd()
+        os.chdir(self.newPath)
+
+    def __exit__(self, etype, value, traceback):
+        os.chdir(self.savedPath)
