+++ test.sh
@@ -476,6 +476,53 @@ function DependencyUpdateTestWithUTF8Character() {
     expect git dependencies update -r
 }
 
+function DependencyCommandSetterTests() {
+  cd project
+  expect git dependencies add '../dependency' dep master
+  git add .
+  git commit -m 'Adding dependency'
+  expect git dependencies update -r
+
+  git dependencies command dep "!sh echo \$(pwd)"
+  local result=$(cat .gitdepends | grep 'command')
+  expect [[ ${#result} -gt 0 ]]
+  expect [[ "\"$result\"" == "\"command = !sh echo \$(pwd)\"" ]]
+
+  git dependencies command dep "!sh echo \$(pwd) > sample.txt && ls"
+  local result2=$(cat .gitdepends | grep 'command')
+  expect [[ ${#result2} -gt 0 ]]
+  expect [[ "\"$result2\"" == "\"command = !sh echo \$(pwd) > sample.txt && ls\"" ]]
+
+  git dependencies command dep ""
+  local result3=$(cat .gitdepends | grep 'command')
+  expect [[ ${#result3} -eq 0 ]]
+}
+
+function DependencyCommandRunnerTests() {
+  cd dependency
+  echo $'#!/bin/bash\n\necho \"Hello World!\"' > sample.sh
+  chmod u+x sample.sh
+  git add .
+  git commit -m "Add shell script"
+
+  cd ../project
+  expect git dependencies add '../dependency' dep master
+  git add .
+  git commit -m 'Adding dependency'
+  expect git dependencies update -r
+
+  git dependencies command dep "!sh sh ./dep/sample.sh > hello.txt"
+  cat .gitdepends
+  git dependencies update
+
+  expect [[ -f hello.txt ]]
+  expect [[ "\"$(cat hello.txt)\"" == "\"Hello World!\"" ]]
+
+  git dependencies command dep "!sh ln -s dep/README README_LINK"
+  git dependencies update
+  expect [[ -L README_LINK ]]
+}
+
 # TODO
 # 1. test submodules
 # 2. test selective commands (with dependency path specified)
